+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/viewmodels/BaseSheetViewModel.kt
@@ -20,6 +20,7 @@ import com.stripe.android.paymentsheet.analytics.EventReporter
 import com.stripe.android.paymentsheet.analytics.PaymentSheetAnalyticsListener
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.navigation.NavigationHandler
+import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
 import com.stripe.android.paymentsheet.state.WalletsProcessingState
 import com.stripe.android.paymentsheet.state.WalletsState
@@ -54,7 +55,10 @@ internal abstract class BaseSheetViewModel(
     private val _paymentMethodMetadata = MutableStateFlow<PaymentMethodMetadata?>(null)
     internal val paymentMethodMetadata: StateFlow<PaymentMethodMetadata?> = _paymentMethodMetadata
 
-    val navigationHandler: NavigationHandler = NavigationHandler(viewModelScope) { poppedScreen ->
+    val navigationHandler: NavigationHandler<PaymentSheetScreen> = NavigationHandler(
+        coroutineScope = viewModelScope,
+        initialScreen = PaymentSheetScreen.Loading,
+    ) { poppedScreen ->
         analyticsListener.reportPaymentSheetHidden(poppedScreen)
     }