+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/PaymentSheetTopBarStateFactoryTest.kt
@@ -1,32 +1,10 @@
 package com.stripe.android.paymentsheet.ui
 
 import com.google.common.truth.Truth.assertThat
-import com.stripe.android.paymentsheet.R
 import org.junit.Test
 import com.stripe.android.R as StripeR
-import com.stripe.android.ui.core.R as StripeUiCoreR
 
 class PaymentSheetTopBarStateFactoryTest {
-
-    @Test
-    fun `navigation is close when canNavigateBack=false`() {
-        val state = buildTopBarState(
-            canNavigateBack = false,
-        )
-
-        assertThat(state.icon).isEqualTo(R.drawable.stripe_ic_paymentsheet_close)
-        assertThat(state.contentDescription).isEqualTo(R.string.stripe_paymentsheet_close)
-    }
-
-    @Test
-    fun `navigation is back when canNavigateBack=true`() {
-        val state = buildTopBarState(
-            canNavigateBack = true,
-        )
-
-        assertThat(state.icon).isEqualTo(R.drawable.stripe_ic_paymentsheet_back)
-        assertThat(state.contentDescription).isEqualTo(StripeUiCoreR.string.stripe_back)
-    }
 
     @Test
     fun `showTestModeLabel=true when isLiveMode=false`() {
@@ -119,12 +97,10 @@ class PaymentSheetTopBarStateFactoryTest {
     }
 
     private fun buildTopBarState(
-        canNavigateBack: Boolean = false,
         isLiveMode: Boolean = false,
         editable: PaymentSheetTopBarState.Editable = PaymentSheetTopBarState.Editable.Never,
     ): PaymentSheetTopBarState {
         return PaymentSheetTopBarStateFactory.create(
-            hasBackStack = canNavigateBack,
             isLiveMode = isLiveMode,
             editable,
         )