+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/verticalmode/DefaultVerticalModeFormInteractorTest.kt
@@ -59,19 +59,6 @@ internal class DefaultVerticalModeFormInteractorTest {
         }
     }
 
-    @Test
-    fun `canGoBack calls delegate`() {
-        var canGoBack = false
-        runScenario(
-            selectedPaymentMethodCode = "card",
-            canGoBackDelegate = { canGoBack },
-        ) {
-            assertThat(interactor.canGoBack()).isFalse()
-            canGoBack = true
-            assertThat(interactor.canGoBack()).isTrue()
-        }
-    }
-
     private val notImplemented: () -> Nothing = { throw AssertionError("Not implemented") }
 
     private fun runScenario(
@@ -80,7 +67,6 @@ internal class DefaultVerticalModeFormInteractorTest {
             notImplemented()
         },
         reportFieldInteraction: (String) -> Unit = { notImplemented() },
-        canGoBackDelegate: () -> Boolean = { notImplemented() },
         testBlock: suspend TestParams.() -> Unit,
     ) {
         val formArguments = mock<FormArguments>()
@@ -96,7 +82,6 @@ internal class DefaultVerticalModeFormInteractorTest {
             reportFieldInteraction = reportFieldInteraction,
             headerInformation = null,
             isLiveMode = true,
-            canGoBackDelegate = canGoBackDelegate,
             processing = processing,
             coroutineScope = CoroutineScope(UnconfinedTestDispatcher()),
             paymentMethodIncentive = stateFlowOf(null),